Consolidate common first layer commonalities between submission and query internal classes.
