--- Sources/BVSwift/BVCommon/Query/BVInternalQueryRequestable.swift
@@ -1,40 +1,0 @@
-//
-//  BVInternalQueryRequestable.swift
-//  BVSwift
-//
-//  Copyright © 2018 Bazaarvoice. All rights reserved.
-//
-
-import Foundation
-
-extension BVInternalQuery: BVURLRequestable {
-  
-  internal var request: URLRequest? {
-    if commonEndpoint.isEmpty {
-      fatalError(
-        "Endpoint value is empty, make sure you configure the query first.")
-    }
-    
-    let urlString: String = "\(commonEndpoint)\(bvPath)"
-    guard var urlComponents: URLComponents =
-      URLComponents(string: urlString) else {
-        return nil
-    }
-    
-    if let items = urlQueryItems,
-      !items.isEmpty {
-      urlComponents.queryItems = items
-    }
-    
-    guard let url: URL = urlComponents.url else {
-      return nil
-    }
-    
-    BVLogger
-      .sharedLogger.debug("Issuing Query Request to: \(url.absoluteString)")
-    
-    let cachePolicy: URLRequest.CachePolicy =
-      usesURLCache ? .returnCacheDataElseLoad : .reloadIgnoringLocalCacheData
-    return URLRequest(url: url, cachePolicy: cachePolicy)
-  }
-}
